+++ lib/context.ts
@@ -542,23 +542,11 @@ export class Context {
 
   /**
    *
-   * @returns {Promise<number>} Promise that resolves into the resulting chat id
+   * @returns resulting chat id or 0 on error
    */
-  joinSecurejoin(qrCode: string): Promise<number> {
+  joinSecurejoin(qrCode: string): number {
     debug(`joinSecurejoin ${qrCode}`)
-    return new Promise((resolve, reject) => {
-      binding.dcn_join_securejoin(
-        this.dcn_context,
-        qrCode,
-        (result: number) => {
-          if (result !== 0) {
-            resolve(result)
-          } else {
-            reject('The out-of-band verification failed or was aborted')
-          }
-        }
-      )
-    })
+    return binding.dcn_join_securejoin(this.dcn_context, qrCode)
   }
 
   lookupContactIdByAddr(addr: string): number {